3주차 과제

+++ src/main/java/org/sopt/assignment/service/PostService.java
@@ -2,10 +2,13 @@ package org.sopt.assignment.service;
 
 import jakarta.persistence.EntityNotFoundException;
 import org.sopt.assignment.domain.Post;
+import org.sopt.assignment.domain.Tag;
+import org.sopt.assignment.domain.User;
 import org.sopt.assignment.dto.response.PostDetailResponse;
 import org.sopt.assignment.dto.response.PostUpdateResponse;
 import org.sopt.assignment.exception.ErrorMessage;
 import org.sopt.assignment.repository.PostRepository;
+import org.sopt.assignment.repository.UserRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -17,9 +20,11 @@ import java.util.List;
 public class PostService {
 
     private final PostRepository postRepository;
+    private final UserRepository userRepository;
 
-    public PostService(PostRepository postRepository) {
+    public PostService(PostRepository postRepository, UserRepository userRepository) {
         this.postRepository = postRepository;
+        this.userRepository = userRepository;
     }
 
     // 하나의 게시글 조회
@@ -30,12 +35,12 @@ public class PostService {
 
     // 게시글 전체 조회
     public List<Post> getAllPosts() {
-        return postRepository.findAll();
+        return postRepository.findAllByOrderByCreatedAtDesc();
     }
 
     // 게시글 작성
     @Transactional
-    public Long createPost(String title) {
+    public Long createPost(String title, String content, Long userId, Tag tag) {
         // 중복 제목 검사
         if (postRepository.existsByTitle(title)) {
             throw new IllegalArgumentException(ErrorMessage.DUPLICATE_TITLE.getMessage());
@@ -49,33 +54,49 @@ public class PostService {
                 throw new IllegalArgumentException(ErrorMessage.SPAM_LIMIT.getMessage());
             }
         }
-        Post saved = postRepository.save(new Post(title));
+
+        //유저 유효성 검사
+        User user = userRepository.findById(userId)
+                .orElseThrow( ()-> new IllegalArgumentException(ErrorMessage.USER_NOT_FOUND.getMessage()));
+
+        Post saved = postRepository.save(new Post(title,content,user,tag));
         return saved.getId();
     }
 
     // 게시글 수정
     @Transactional
-    public PostUpdateResponse updatePostTitle(Long id, String newTitle) {
-        Post post = getPostById(id);
-        post.updateTitle(newTitle);
-        return new PostUpdateResponse(post.getId(), post.getTitle());
+    public PostUpdateResponse updatePost(Long postId, Long userId, String title, String content, Tag tag) {
+        Post post = getPostById(postId);
+
+        if (!post.getUser().getId().equals(userId)) {
+            throw new IllegalArgumentException(ErrorMessage.UNAUTHORIZED_USER.getMessage());
+        }
+
+        post.update(title, content, tag);
+        return new PostUpdateResponse(post.getId(), post.getTitle(), post.getContent(), post.getTag());
     }
 
     // 게시글 상세 조회
     @Transactional(readOnly = true)
     public PostDetailResponse getPostDetail(Long id) {
         Post post = getPostById(id);
-        return new PostDetailResponse(post.getId(), post.getTitle());
+        return new PostDetailResponse(post.getId(), post.getTitle(),post.getContent(),post.getUser().getNickname(),post.getTag());
     }
 
     //게시글 삭제
+
     @Transactional
-    public void deletePostById(Long id) {
-        Post post = getPostById(id);
+    public void deletePostById(Long postId, Long userId) {
+        Post post = getPostById(postId);
+
+        if (!post.getUser().getId().equals(userId)) {
+            throw new IllegalArgumentException(ErrorMessage.UNAUTHORIZED_USER.getMessage());
+        }
+
         postRepository.delete(post);
     }
 
-    //게시글 검색
+    //게시글 검색 (제목 기준)
     @Transactional(readOnly = true)
     public List<Post> searchPostsByKeyword(String keyword) {
         List<Post> result = postRepository.searchByKeyword(keyword);
@@ -89,4 +110,23 @@ public class PostService {
         return result;
     }
 
+    //게시글 검색 (작성자 기준)
+    @Transactional(readOnly = true)
+    public List<Post> searchPostsByWriter(String nickname) {
+        List<Post> result = postRepository.searchByWriter(nickname);
+
+        if (result.isEmpty()) {
+            throw new EntityNotFoundException(ErrorMessage.WRITER_NOT_FOUND.getMessage());
+        }
+        return result;
+    }
+
+    @Transactional(readOnly = true)
+    public List<Post> searchByTag(Tag tag) {
+        List<Post> result = postRepository.findByTag(tag);
+        if (result.isEmpty()) {
+            throw new EntityNotFoundException(ErrorMessage.TAG_NOT_FOUND.getMessage());
+        }
+        return result;
+    }
 }
